Move child process cleanup to the main function
That way it's alongside the code that created the child process in the first
place.

The contract is that the child should exit if it fails to initialize, so
there's no need to send it SIGTERM.

The child can be told to exit by shutting down the parent end of the parent/child
socket.

The child no longer needs to handle SIGTERM any more, so remove that code.