Fix deadlock when parent sends exactly SOCKET_BUFFER_LEN bytes to the child
The parent now sends the length of the picked data before sending the
data itself. This means the client knows how much data to expect, and
can therefore read until it recieves that much, no more, no less.

At the same time, we can simplify communication between the child's main
thread. There's no need to go to the length of transmitting the pickled
data via an internal socketpair; both threads are in the same process so
provided the right locking is done (via a condition variable in this
case), they can share the recieved data directly.