spec: add missing BuildRequries on systemd-rpm-macros
When rpm ran the scriptlets, they failed with e.g.,

  /var/tmp/rpm-tmp.gGJWcJ: line 1: fg: no job control
  error: %preun(certmonger-exporter-0.1-1.git.86.8282852.el9.noarch) scriptlet failed, exit status 1

  Error in PREUN scriptlet in rpm package certmonger-exporter

This is a tricky one. At firt glance the scripts look OK:

  $ rpm -q certmonger-exporter --scripts
  postinstall scriptlet (using /bin/sh):
  %systemd_post certmonger-exporter.service
  preuninstall scriptlet (using /bin/sh):
  %systemd_preun certmonger-exporter.service
  postuninstall scriptlet (using /bin/sh):
  %systemd_postun_with_restart certmonger-exporter.service

But actaully these macros are supposed to have been expanded at build
time! So when the shell tried to run the scripts, it interpreted the
pipelines starting with % as job-control directives.

The fix is to add systemd-rpm-macros to BuildRequires so that the macros
are expanded at build time.

This _is_ picked up on by rpmlint:

  certmonger-exporter.noarch: W: percent-in-%preun
  certmonger-exporter.noarch: W: percent-in-%postun
  certmonger-exporter.noarch: W: percent-in-%post

... so I guess fool me for not noticing.